% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_natural.R
\name{arrange_natural}
\alias{arrange_natural}
\title{Arrange rows by column values naturally}
\usage{
arrange_natural(.data, ...)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble), or a lazy
data frame (e.g. from dbplyr or dtplyr).}

\item{...}{<\code{\link[=dplyr_data_masking]{data-masking}}> Variables to sort by.}
}
\value{
An object of the same type as \code{.data}. The output has the following
properties:
\itemize{
\item All rows appear in the output, but (usually) in a different place.
\item Columns are not modified.
\item Data frame attributes are preserved.
}
}
\description{
\code{arrange_natural()} orders the rows of a data frame by the values of selected
columns.

Data is naturally sorted (see \emph{Details}, below) in ascending order.
Grouping is ignored.
}
\details{
\subsection{Natural sorting}{

\code{arrange_natural()} is built on \code{\link[dplyr:arrange]{dplyr::arrange()}} to provide
\href{https://en.wikipedia.org/wiki/Natural_sort_order}{natural sorting}
(sorting of strings with both letters and numerals). The underlying
implementation for natural sorting is based on the
\href{https://stringi.gagolewski.com/}{\code{stringi}} library.
}
}
\examples{
df <- tibble::tribble(
  ~sample, ~gene,
  "D10-23", "atp6",
  "D10-43", "mdr1",
  "D10-55", "atp6",
  "D10-5", "mdr1",
  "D10-47", "dhps",
  "D10-15", "atp6"
)

arrange_natural(df, sample)
df \%>\% arrange_natural(sample, gene)
}
\seealso{
\code{\link[dplyr:arrange]{dplyr::arrange()}}
}
