% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromosome-map.R
\name{chromosome-map}
\alias{chromosome-map}
\alias{plot_chromoMap}
\alias{plot_karyoploteR}
\title{Create annotated chromosome map}
\usage{
plot_chromoMap(genome, probes, title = "", colours = list(), ...)

plot_karyoploteR(genome, probes, title = "", colours = list(), ...)
}
\arguments{
\item{genome}{A tibble indicating the starting and ending position of each
chromosome. Contains three columns:
\itemize{
\item Name of the chromosome
\item The starting position of the chromosome
\item The ending position of the chromosome.
}}

\item{probes}{A tibble indicating the starting and ending position of each
probe. Contains four columns:
\itemize{
\item Name of the chromosome the probe is on
\item The starting position of the probe
\item The ending position of the probe
\item An identifier indicating the probe set the probe belongs to.
}}

\item{title}{The title of the plot.}

\item{colours}{A vector of colours indicating the annotation colour for each
probe set.}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Additional arguments passed to
internal plotting functions.}
}
\value{
A \link[chromoMap:chromoMap]{chromoMap} or
\href{https://bernatgel.github.io/karyoploter_tutorial/}{karyoploteR} object.
}
\description{
Render a graphics visualization of entire chromosomes or chromosomal regions.
Annotate multiple targeted regions to visualize probe targets.
}
\examples{
\donttest{
probes <- tibble::tribble(
  ~chrom, ~start, ~end, ~probe_set,
  "chr14", 2342135L, 2342284L, "IBC",
  "chr3", 830503L, 830769L, "DR2",
  "chr5", 482233L, 482391L, "IBC",
  "chr9", 375274L, 375417L, "IBC",
  "chr12", 532032L, 532281L, "DR2",
  "chr7", 383447L, 383653L, "HAP",
  "chr14", 1401991L, 1402160L, "IBC",
  "chr4", 734737L, 734936L, "HAP",
  "chr10", 93054L, 93223L, "IBC",
  "chr7", 162127L, 162277L, "IBC"
)

plot_chromoMap(genome_Pf3D7, probes)

plot_karyoploteR(
  genome_Pf3D7,
  probes,
  title = "Example Chromosome Map",
  colours = c("#006A8EFF", "#A8A6A7FF", "#B1283AFF")
)}
}
\seealso{
See the \link[chromoMap:chromoMap]{chromoMap} and
\href{https://bernatgel.github.io/karyoploter_tutorial/}{karyoploteR} packages
to create custom chromosome maps.
}
