% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_single}
\alias{convert_single}
\alias{convert}
\alias{convert_three}
\title{Convert amino acid abbreviations}
\usage{
convert_single(str)

convert_three(str)
}
\arguments{
\item{str}{Character vector containing amino acids to convert.}
}
\value{
Character vector with converted amino acid abbreviations.
}
\description{
\code{convert_single()} converts the 1-letter amino acid abbreviation to the
3-letter amino acid abbreviation and \code{convert_three()} does the opposite.
}
\details{
Conversion is case-insensitive, but always returns capitalized 1-letter and
3-letter abbreviations in title-case.
}
\examples{
# Convert 1-letter abbreviation
convert_single("G235L")
convert_single("s545p")

# Convert 3-letter abbreviation
convert_three("Gly235Leu")
convert_three("SER545LEU")
}
