% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{custom_themes}
\alias{theme_miplicorn}
\alias{theme_rainbow}
\title{Custom themes}
\usage{
theme_miplicorn(
  base_size = 10,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)

theme_rainbow(
  base_size = 12,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{Base font size, given in pts.}

\item{base_family}{Base font family.}

\item{base_line_size}{Base size for line elements.}

\item{base_rect_size}{Base size for rect elements.}
}
\value{
A \link[ggplot2:theme]{ggplot2} theme.
}
\description{
These are custom \code{ggplot2} themes which control all non-data display. They
may further tweak components of these themes using \code{\link[ggplot2:theme]{ggplot2::theme()}}.
}
\examples{
data <- tibble::tibble(
  gene = c("atp6", "k13", "dhps"),
  coverage = c(10, 100, 50)
)

p <- ggplot2::ggplot(data, ggplot2::aes(x = gene, y = coverage, fill = factor(gene))) +
  ggplot2::geom_col()

p + theme_miplicorn()
p + theme_rainbow()
}
