% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter}
\alias{filter}
\alias{filter_aa_change}
\alias{filter_alt_umi_count}
\alias{filter_coverage}
\alias{filter_gene}
\alias{filter_mutation_name}
\alias{filter_ref_umi_count}
\alias{filter_targeted}
\title{Subset rows using a single column value}
\usage{
filter_aa_change(.data, .value, .preserve = FALSE)

filter_alt_umi_count(.data, .value, .preserve = FALSE)

filter_coverage(.data, .value, .preserve = FALSE)

filter_gene(.data, .value, .preserve = FALSE)

filter_mutation_name(.data, .value, .preserve = FALSE)

filter_ref_umi_count(.data, .value, .preserve = FALSE)

filter_targeted(.data, .value, .preserve = FALSE)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble), or a lazy
data frame (e.g. from dbplyr or dtplyr).}

\item{.value}{Filtering value. Data greater than or equal to the value will
be kept.}

\item{.preserve}{Relevant when the \code{.data} input is grouped.
If \code{.preserve = FALSE} (the default), the grouping structure
is recalculated based on the resulting data, otherwise the grouping is kept as is.}
}
\description{
These \verb{filter_*()} functions are used to subset a data frame. Each function
subsets rows based on the value of one column. For numeric columns, a row
will be kept if the value is greater than or equal to the value specified.
For categorical columns, a row will be kept if the value is equal to the
value specified. Note that when a condition evaluates to NA the row will be
dropped, unlike base subsetting with \code{[}.
}
\examples{
data <- tibble::tribble(
  ~sample, ~gene, ~coverage,
  "S1", "atp6", 10,
  "S2", "crt", 20,
)
filter_gene(data, "atp6")
filter_coverage(data, 15)
}
\seealso{
\code{\link[dplyr:filter]{dplyr::filter()}} for more complex filtering operations.
}
