% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_fn_factory}
\alias{filter_fn_factory}
\title{Filter function factory}
\usage{
filter_fn_factory(.col, .type = c("numeric", "character"))
}
\arguments{
\item{.col}{The name of the column used to filter the data.}

\item{.type}{The type of the column.}
}
\description{
Define a function factory to create filtering functions that filter the
data based on a single column.
}
\examples{
\dontrun{

filter_coverage <- filter_fn_factory("coverage", "numeric")
filter_gene <- filter_fn_factory("gene", "character")
}
}
\seealso{
\href{https://adv-r.hadley.nz/function-factories.html}{Advanced R} for
more information on function factories.
}
