% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-genomes.R
\docType{data}
\name{genome_Pf3D7}
\alias{genome_Pf3D7}
\title{Plasmodium falciparum 3D7 genome}
\format{
A tibble with 16 rows and 3 variables:
\describe{
\item{chrom}{Name of the chromosome}
\item{start}{The starting position of the chromosome}
\item{end}{The ending position of the chromosome}
}
}
\source{
\href{https://plasmodb.org/plasmo/app/downloads/Current_Release/Pfalciparum3D7/gff/data/PlasmoDB-54_Pfalciparum3D7.gff}{PlasmoDB Pf. 3D7}
}
\usage{
genome_Pf3D7
}
\description{
This data set contains information on the P. falciparum 3D7 reference genome.
The original source is from \href{https://plasmodb.org/plasmo/app/}{PlasmoDB}.
}
\examples{
genome_Pf3D7
}
\keyword{datasets}
