% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_mutations.R
\name{label_mutations}
\alias{label_mutations}
\title{Label mutations}
\usage{
label_mutations(.data, .before = NULL, .after = NULL)
}
\arguments{
\item{.data}{The data set containing REF and ALT calls.}

\item{.before, .after}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
<\code{\link[=dplyr_tidy_select]{tidy-select}}> Optionally, control where new columns
should appear (the default is to add to the right hand side). See
\code{\link[=relocate]{relocate()}} for more details.}
}
\description{
For each data point, label a mutation as \code{"ref"}, \code{"alt"}, \code{"ins"}, or
\code{"del"}.
\itemize{
\item Data points that have more reference (REF) than alternate (ALT) calls will
be labeled as \code{"ref"}.
\item Mutations whose REF and ALT calls are both one base pair and who have more
ALT calls than REF calls will be labeled as \code{"alt"}.
\item Mutations whose REF and ALT calls differ in length and have more ALT calls
than REF calls will be labeled as \code{"ins"}.
\item Mutations whose REF and ALT calls differ in length and have more REF calls
than ALT calls will be labeled as \code{"del"}.
}
}
\examples{
data <- tibble::tribble(
  ~sample, ~pos, ~ref, ~alt, ~ref_umi_count, ~alt_umi_count, ~coverage,
  "S1", "1049838", "A", "G", 54, 10, 64,
  "S2", "801498", "G", "A", 15, 0, 15,
  "S3", "625403", "T", "C", 0, 15, 15,
  "S4", "748165", "GA", "G", 2, 18, 20,
  "S5", "487199", "G", "ATC", 0, 10, 10
)

label_mutations(data)
label_mutations(data, .after = alt)
label_mutations(data, .before = pos)
}
