% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{miplicorn_example}
\alias{miplicorn_example}
\alias{example}
\title{Get path to miplicorn examples}
\usage{
miplicorn_example(path = NULL)
}
\arguments{
\item{path}{Name of file. If \code{NULL}, all example files will be listed.}
}
\value{
A local filepath or a list of all available files.
}
\description{
miplicorn contains several example files in its 'inst/extdata' directory.
This function can be used to access file paths.
}
\examples{
# Get path to one example
miplicorn_example("reference_AA_table.csv")

# List all available examples
miplicorn_example()
}
