% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIPr.R
\docType{package}
\name{MIPr}
\alias{MIPr}
\alias{MIPr-package}
\title{\pkg{MIPr} Collection of Functions for MIPTools Analysis}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Aris Paschalidis \email{aris.paschalidis@gmail.com} (\href{https://orcid.org/0000-0003-2247-1885}{ORCID})

}
