% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{MIPr_example}
\alias{MIPr_example}
\alias{example}
\alias{examples}
\alias{MIPr_examples}
\title{Get path to MIPr examples}
\usage{
MIPr_example(path)

MIPr_examples(pattern = NULL)
}
\arguments{
\item{path}{Name of file.}

\item{pattern}{A regular expression of filenames to match. If \code{NULL} all
available files are listed.}
}
\description{
MIPr contains several example files in its 'inst/extdata' directory. Use
\code{MIPr_example()} to retrieve the path to one file and \code{MIPr_examples()} to
list all the available files.
}
\details{
Thanks to Jim Hester's and Hadley Wickham's \code{\link[vroom]{vroom}}
package for these functions.
}
\examples{
# Get path to one example
MIPr_example("reference_AA_table.csv")

# List all available examples
MIPr_examples()
}
