% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation-frequency.R
\name{mutation_frequency}
\alias{mutation_frequency}
\alias{mutation_frequency.ref_alt_cov_tbl}
\title{Compute frequency of mutations}
\usage{
mutation_frequency(data, threshold)

\method{mutation_frequency}{ref_alt_cov_tbl}(data, threshold)
}
\arguments{
\item{data}{A data frame, data frame extension (e.g. a tibble), or a lazy
data frame (e.g. from dbplyr or dtplyr).}

\item{threshold}{A minimum UMI count which reflects the confidence in the
genotype call. Data with a UMI count of less than the threshold will be
filtered out from the analysis.}
}
\value{
A \link[tibble:tibble-package]{tibble} with the extra class \code{mut_freq}. The
output has the following columns:
\itemize{
\item \code{mutation_name}: The unique mutation sequenced.
\item \code{frequency}: The frequency of the mutation.
}
}
\description{
Generate a table representing the frequency of unique mutations. In order to
ensure confidence in the results, a threshold is provided indicating
confidence in genotype calls. All data that do not meet this threshold will
be removed from the computation.
}
\examples{
ref_file <- miplicorn_example("reference_AA_table.csv")
alt_file <- miplicorn_example("alternate_AA_table.csv")
cov_file <- miplicorn_example("coverage_AA_table.csv")
data <- read_tbl_ref_alt_cov(
  ref_file,
  alt_file,
  cov_file,
  gene == "atp6" | gene == "crt"
)
mutation_frequency(data, 5)
}
\seealso{
\code{\link[=plot_mutation_frequency]{plot_mutation_frequency()}} for plotting the table.
}
