% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coverage.R
\name{plot_coverage}
\alias{plot_coverage}
\alias{plot_coverage.cov_tbl}
\alias{plot_coverage.ref_alt_cov_tbl}
\title{Plot coverage}
\usage{
plot_coverage(data, group_by)

\method{plot_coverage}{cov_tbl}(data, group_by)

\method{plot_coverage}{ref_alt_cov_tbl}(data, group_by)
}
\arguments{
\item{data}{A data frame, data frame extension (e.g. a tibble), or a lazy
data frame (e.g. from dbplyr or dtplyr).}

\item{group_by}{The grouping variable. The average coverage is computed for
each unique value of the variable.}
}
\value{
A \link[ggplot2:ggplot2-package]{ggplot2} object.
}
\description{
Create a bar chart visualizing the average coverage across a given grouping
variable. The color of each bar represents the magnitude of the mean
coverage.
}
\examples{
# Read example data
data <- read_tbl_coverage(miplicorn_example("coverage_AA_table.csv"))

# Plot coverage grouped by gene
plot_coverage(data, gene)
}
