% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainbow_haplotype.R
\name{plot_haplotypes}
\alias{plot_haplotypes}
\title{Create rainbow haplotype plot}
\usage{
plot_haplotypes(
  data,
  sample = s_Sample,
  probe = p_name,
  haplotype = h_popUID,
  rel_freq = c_AveragedFrac,
  n_colours = 11,
  ...
)
}
\arguments{
\item{data}{The data to be plotted.}

\item{sample}{The column name representing the samples.}

\item{probe}{The column name representing the probes.}

\item{haplotype}{The column name representing the haplotypes.}

\item{rel_freq}{The column name representing relative abundance or frequency
of each haplotype.}

\item{n_colours}{An integer indicating the number of colors to be used.}

\item{...}{These dots are for future extensions and must be empty.}
}
\description{
Create a rainbow haplotype plot. Each row represents a sample and each column
represents a unique probe. The colors in each bar represent the haplotypes
identified for each probe. The height of the bar illustrates the relative
abundance of each haplotype.
}
\seealso{
\code{\link[=prep_haplotypes]{prep_haplotypes()}} for generating the data used for plotting.
}
