% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation-frequency.R
\name{plot_mutation_frequency}
\alias{plot_mutation_frequency}
\alias{autoplot.mut_freq}
\alias{plot.mut_freq}
\title{Plot frequency of mutations}
\usage{
plot_mutation_frequency(data)

\method{autoplot}{mut_freq}(object, ...)

\method{plot}{mut_freq}(x, ...)
}
\arguments{
\item{data, object, x}{An object of class \code{mut_freq}. Derived from the output
of \code{\link[=mutation_frequency]{mutation_frequency()}}.}

\item{...}{Other arguments passed to specific methods.}
}
\value{
A \link[ggplot2:ggplot2-package]{ggplot2} object.
}
\description{
Plot the frequency of mutations generated by \code{\link[=mutation_frequency]{mutation_frequency()}}.
The frequency is plotted on the y-axis and the amino acid change is plotted
on the x-axis. Data are grouped by the gene on which the mutation took place
and coloured according to their groupings.
}
\examples{
# Read example data
data <- read_tbl_ref_alt_cov(
  miplicorn_example("reference_AA_table.csv"),
  miplicorn_example("alternate_AA_table.csv"),
  miplicorn_example("coverage_AA_table.csv"),
  gene == "atp6" | gene == "crt"
)

# Compute the mutation frequency
frequency <- mutation_frequency(data, 5)

# Plot
plot(frequency)
}
\seealso{
\code{\link[=mutation_frequency]{mutation_frequency()}} for generating the data for plotting.
}
