% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation-prevalence.R
\name{plot_mutation_prevalence}
\alias{plot_mutation_prevalence}
\title{Plot prevalence of mutations}
\usage{
plot_mutation_prevalence(data)
}
\arguments{
\item{data}{An object of class \code{mutation_prev}. Derived from the output of
\code{\link[=mutation_prevalence]{mutation_prevalence()}}.}
}
\description{
Plot the prevalence of mutations generated by \code{\link[=mutation_prevalence]{mutation_prevalence()}}.
The prevalence is plotted on the y-axis and the amino acid change is plotted
on the x-axis. Data are grouped by the gene on which the mutation took place
and coloured according to their groupings.
}
\examples{
ref_file <- miplicorn_example("reference_AA_table.csv")
alt_file <- miplicorn_example("alternate_AA_table.csv")
cov_file <- miplicorn_example("coverage_AA_table.csv")
data <- read_tbl_ref_alt_cov(
  ref_file,
  alt_file,
  cov_file,
  gene == "atp6" | gene == "crt"
)
prevalence <- mutation_prevalence(data, 5)
plot_mutation_prevalence(prevalence)
}
\seealso{
\code{\link[=mutation_prevalence]{mutation_prevalence()}} for generating the data for plotting.
}
