% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainbow_haplotype.R
\name{prep_haplotypes}
\alias{prep_haplotypes}
\title{Prepare data for rainbow haplotype plot}
\usage{
prep_haplotypes(
  data,
  sample = s_Sample,
  probe = p_name,
  haplotype = h_popUID,
  rel_freq = c_AveragedFrac,
  minPopSize = 3,
  n_colours = 11,
  barHeight = 0.8,
  ...
)
}
\arguments{
\item{data}{The data to be plotted.}

\item{sample}{The column name representing the samples.}

\item{probe}{The column name representing the probes.}

\item{haplotype}{The column name representing the haplotypes.}

\item{rel_freq}{The column name representing relative abundance or frequency
of each haplotype.}

\item{minPopSize}{The minimum number of haplotypes required per probe. If
there are less haplotypes, the probe will be ignored.}

\item{n_colours}{An integer indicating the number of colors to be used.}

\item{barHeight}{The height of the bar used for plotting.}

\item{...}{These dots are for future extensions and must be empty.}
}
\description{
Prepare data for a rainbow haplotype plot. Creates several additional
columns, adding them to the dataset.
}
\seealso{
\code{\link[=plot_haplotypes]{plot_haplotypes()}} for creating the rainbow haplotype plot.
}
