% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read}
\alias{read}
\title{Read data}
\usage{
read(alternate_table, coverage_table)
}
\arguments{
\item{alternate_table}{Alternate table.}

\item{coverage_table}{Coverage table.}
}
\value{
A tibble containing the parsed data.
}
\description{
Read paired data files and combine the files.
}
