% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_file}
\alias{read_file}
\title{Read file}
\usage{
read_file(file, type = c("alternate", "coverage"))
}
\arguments{
\item{file}{Alternate table.}
}
\value{
Parsed file.
}
\description{
Read data files.
}
\keyword{internal}
