% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_table}
\alias{read_table}
\title{Read file}
\usage{
read_table(file, type = c("alternate", "coverage"))
}
\arguments{
\item{file}{The path to the file to be read.}

\item{type}{Whether the table is an alternate or coverage table.}
}
\value{
Parsed file.
}
\description{
Read data table.
}
