% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort.R
\name{sort}
\alias{sort}
\title{Sort rows by column values}
\usage{
sort(.data, ...)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble), or a lazy
data frame (e.g. from dbplyr or dtplyr).}

\item{...}{<\code{\link[=dplyr_data_masking]{data-masking}}> Variables to sort by.}
}
\value{
An object of the same type as \code{.data}. The output has the following
properties:
\itemize{
\item All rows appear in the output, but (usually) in a different place.
\item Columns are not modified.
\item Data frame attributes are preserved.
}
}
\description{
\code{sort()} orders the rows of a data frame by the values of selected columns.
}
\details{
\subsection{Ordering}{

Data is sorted in ascending order.
}

\subsection{Natural sorting}{

\code{sort()} is built on \code{dplyr::arrange()} to provide
\href{https://en.wikipedia.org/wiki/Natural_sort_order}{natural sorting}
(sorting of strings with both letters and numerals). The underlying
implementation for natural sorting is based on the \code{stringi} library.
}
}
\examples{
df <- tibble::tribble(
  ~sample, ~gene,
  "D10-23", "atp6",
  "D10-43", "mdr1",
  "D10-55", "atp6",
  "D10-5", "mdr1",
  "D10-47", "dhps",
  "D10-15", "atp6"
)

sort(df, sample, gene)
df \%>\% sort(sample, gene)
}
\seealso{
\code{\link[dplyr:arrange]{dplyr::arrange()}}
}
