% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_single}
\alias{convert_single}
\alias{convert}
\alias{convert_three}
\title{Convert amino acid abbreviation}
\usage{
convert_single(str)

convert_three(str)
}
\arguments{
\item{str}{String containing amino acids to convert.}
}
\description{
\code{convert_single()} converts the 1-letter abbreviation to the 3-letter
abbreviation and \code{convert_three()} does the opposite.
}
\details{
Conversion is case-insensitive, but always returns capitalized 1-letter
abbreviation and 3-letter abbreviation in title-case.
}
\examples{
# Convert 1-letter abbreviation
convert_single("G235L")
convert_single("s545p")

# Convert 3-letter abbreviation
convert_three("Gly235Leu")
convert_three("SER545LEU")
}
