% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{custom_themes}
\alias{theme_miplicorn}
\alias{theme_rainbow}
\title{Custom themes}
\usage{
theme_miplicorn(
  base_size = 10,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)

theme_rainbow(
  base_size = 12,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{Base font size, given in pts.}

\item{base_family}{Base font family.}

\item{base_line_size}{Base size for line elements.}

\item{base_rect_size}{Base size for rect elements.}
}
\description{
These are custom \code{ggplot2} themes which control all non-data display. You
may further tweak components of these themes using \code{\link[ggplot2:theme]{ggplot2::theme()}}.
}
\examples{
library("ggplot2")
p <- ggplot(mtcars, aes(x = wt, y = mpg, colour = factor(gear))) +
  geom_point() +
  facet_wrap(~am) +
  geom_smooth(method = "lm", se = FALSE)

p + theme_miplicorn()
p + theme_rainbow()
}
