% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miplicorn-package.R
\docType{package}
\name{miplicorn-package}
\alias{miplicorn}
\alias{miplicorn-package}
\title{miplicorn: A Framework for MIP and Amplicon Analysis}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A unified framework for molecular inversion probe and amplicon analysis. Provides the tools to parse, manipulate, analyze, and visualize data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bailey-lab.github.io/miplicorn/}
  \item \url{https://github.com/bailey-lab/miplicorn}
  \item Report bugs at \url{https://github.com/bailey-lab/miplicorn/issues}
}

}
\author{
\strong{Maintainer}: Aris Paschalidis \email{aris.paschalidis@gmail.com} (\href{https://orcid.org/0000-0003-2247-1885}{ORCID})

}
\keyword{internal}
