% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{miplicorn_example}
\alias{miplicorn_example}
\alias{example}
\alias{examples}
\alias{miplicorn_examples}
\title{Get path to miplicorn examples}
\usage{
miplicorn_example(path)

miplicorn_examples(pattern = NULL)
}
\arguments{
\item{path}{Name of file.}

\item{pattern}{A regular expression of filenames to match. If \code{NULL} all
available files are listed.}
}
\description{
miplicorn contains several example files in its 'inst/extdata' directory. Use
\code{miplicorn_example()} to retrieve the path to one file and
\code{miplicorn_examples()} to list all the available files.
}
\details{
Thanks to Jim Hester's and Hadley Wickham's \code{\link[vroom]{vroom}}
package for these functions.
}
\examples{
# Get path to one example
miplicorn_example("reference_AA_table.csv")

# List all available examples
miplicorn_examples()
}
