% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation-prevalence.R
\name{plot_mutation_prevalence}
\alias{plot_mutation_prevalence}
\title{Plot prevalence of mutations}
\usage{
plot_mutation_prevalence(data)
}
\arguments{
\item{data}{An object of class \code{mutation_prev}. Derived from the output of
\code{mutation_prevalence()}.}
}
\description{
Plot the prevalence of mutations generated by \code{mutation_prevalence()}.
The prevalence is plotted on the y-axis and the amino acid change is plotted
on the x-axis. Data are grouped by the gene on which the mutation took place
and coloured according to their groupings.
}
