% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation-prevalence.R
\name{plot_mutation_prevalence}
\alias{plot_mutation_prevalence}
\alias{autoplot.mut_prev}
\alias{plot.mut_prev}
\title{Plot prevalence of mutations}
\usage{
plot_mutation_prevalence(data)

\method{autoplot}{mut_prev}(object, ...)

\method{plot}{mut_prev}(x, ...)
}
\arguments{
\item{data, object, x}{An object of class \code{mut_prev}. Derived from the output
of \code{\link[=mutation_prevalence]{mutation_prevalence()}}.}

\item{...}{Other arguments passed to specific methods.}
}
\description{
Plot the prevalence of mutations generated by \code{\link[=mutation_prevalence]{mutation_prevalence()}}.
The prevalence is plotted on the y-axis and the amino acid change is plotted
on the x-axis. Data are grouped by the gene on which the mutation took place
and coloured according to their groupings.
}
\examples{
ref_file <- miplicorn_example("reference_AA_table.csv")
alt_file <- miplicorn_example("alternate_AA_table.csv")
cov_file <- miplicorn_example("coverage_AA_table.csv")
data <- read_tbl_ref_alt_cov(
  ref_file,
  alt_file,
  cov_file,
  gene == "atp6" | gene == "crt"
)
prevalence <- mutation_prevalence(data, threshold = 5)
plot(prevalence)
}
\seealso{
\code{\link[=mutation_prevalence]{mutation_prevalence()}} for generating the data for plotting.
}
