% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read}
\alias{read}
\title{Read data}
\usage{
read(reference_table, alternate_table, coverage_table, chrom = NULL)
}
\arguments{
\item{reference_table}{File path to the reference table.}

\item{alternate_table}{File path to the alternate table.}

\item{coverage_table}{File path to the coverage table.}

\item{chrom}{The chromosome(s) to filter to.}
}
\value{
A tibble containing the parsed data.
}
\description{
Read files containing MIPtools' data tables and combine them.
}
