% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_file}
\alias{read_file}
\title{Read file}
\usage{
read_file(file, type = c("reference", "alternate", "coverage"))
}
\arguments{
\item{file}{The path to the file to be read.}

\item{type}{Whether the table is an alternate or coverage table.}
}
\value{
Parsed file.
}
\description{
Read file containing MIPtools' data table.
}
\keyword{internal}
