% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_file}
\alias{read_file}
\title{Read file}
\usage{
read_file(file, chrom, name = c("ref_umi_count", "alt_umi_count", "coverage"))
}
\arguments{
\item{file}{The path to the file to be read.}

\item{chrom}{The chromosome(s) to filter to.}

\item{name}{The name of the value we are interested in.}
}
\value{
Parsed file.
}
\description{
Read file containing MIPtools' data table.
}
\keyword{internal}
