% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{theme_miplicorn}
\alias{theme_miplicorn}
\alias{theme_rainbow}
\title{Custom ggplot2 themes}
\usage{
theme_miplicorn(
  base_size = 10,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)

theme_rainbow(
  base_size = 12,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}
}
\description{
Custom ggplot2 themes
}
\examples{
library("ggplot2")
p <- ggplot(mtcars, aes(x = wt, y = mpg, colour = factor(gear))) +
  geom_point() +
  facet_wrap(~am) +
  geom_smooth(method = "lm", se = FALSE)

p + theme_miplicorn()
p + theme_rainbow()
}
